# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "dto.CommentVO": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "courseId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "like": {
                        "type": "boolean"
                    },
                    "likeCnt": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "teachers": {
                        "description": "这里还是直接返回教师名字+职称组合后的字符串列表",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.CreateCommentReq": {
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "courseId": {
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "required": [
                    "content",
                    "courseId"
                ],
                "type": "object"
            },
            "dto.CreateCommentResp": {
                "description": "实际业务数据",
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "courseId": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "like": {
                        "type": "boolean"
                    },
                    "likeCnt": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "teachers": {
                        "description": "这里还是直接返回教师名字+职称组合后的字符串列表",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "dto.GetCourseCampusesResp": {
                "description": "实际业务数据",
                "properties": {
                    "campuses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.GetCourseCategoriesResp": {
                "description": "实际业务数据",
                "properties": {
                    "categories": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.GetCourseDepartmentsResp": {
                "description": "实际业务数据",
                "properties": {
                    "departments": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "dto.GetCourseResp": {
                "description": "实际业务数据",
                "properties": {
                    "course": {
                        "$ref": "#/components/schemas/dto.CourseVO"
                    }
                },
                "type": "object"
            },
            "dto.GetMyCommentsReq": {
                "properties": {
                    "page": {
                        "form": "page",
                        "type": "integer"
                    },
                    "pageSize": {
                        "form": "pageSize",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "dto.GetMyCommentsResp": {
                "description": "实际业务数据",
                "properties": {
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/dto.CommentVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "dto.GetTotalCourseCommentsCountResp": {
                "description": "实际业务数据",
                "properties": {
                    "count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "dto.ListCourseCommentsResp": {
                "description": "实际业务数据",
                "properties": {
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/dto.CommentVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "dto.ListCoursesReq": {
                "properties": {
                    "page": {
                        "form": "page",
                        "type": "integer"
                    },
                    "pageSize": {
                        "form": "pageSize",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "dto.ListCoursesResp": {
                "description": "实际业务数据",
                "properties": {
                    "courses": {
                        "description": "当前页的课程列表",
                        "items": {
                            "$ref": "#/components/schemas/dto.CourseVO"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "page": {
                        "form": "page",
                        "type": "integer"
                    },
                    "pageSize": {
                        "form": "pageSize",
                        "type": "integer"
                    },
                    "total": {
                        "description": "符合条件的总记录数",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_CreateCommentResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.CreateCommentResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_GetCourseCampusesResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetCourseCampusesResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetCourseCategoriesResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetCourseCategoriesResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetCourseDepartmentsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetCourseDepartmentsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetCourseResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetCourseResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_GetMyCommentsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetMyCommentsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_GetTotalCourseCommentsCountResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.GetTotalCourseCommentsCountResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "handler.Response-dto_ListCourseCommentsResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListCourseCommentsResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "handler.Response-dto_ListCoursesResp": {
                "properties": {
                    "code": {
                        "description": "业务代码, 0表示成功",
                        "example": 0,
                        "type": "integer"
                    },
                    "data": {
                        "$ref": "#/components/schemas/dto.ListCoursesResp"
                    },
                    "msg": {
                        "description": "提示信息",
                        "example": "success",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "选课猫后端接口文档",
        "title": "Meowpick Backend API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/comment/add": {
            "post": {
                "description": "用户对指定课程发布评论",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.CreateCommentReq",
                                        "summary": "body",
                                        "description": "CreateCommentReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "CreateCommentReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_CreateCommentResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "发布课程评论",
                "tags": [
                    "comment"
                ]
            }
        },
        "/api/comment/history": {
            "post": {
                "description": "分页获取当前用户发布过的评论",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.GetMyCommentsReq",
                                        "summary": "body",
                                        "description": "GetMyCommentsReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "GetMyCommentsReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetMyCommentsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取我的评论历史",
                "tags": [
                    "comment"
                ]
            }
        },
        "/api/comment/query": {
            "get": {
                "description": "根据课程ID分页查询评论列表",
                "parameters": [
                    {
                        "description": "课程ID",
                        "in": "query",
                        "name": "courseId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "页码",
                        "in": "query",
                        "name": "page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "in": "query",
                        "name": "pageSize",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListCourseCommentsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "分页获取课程评论",
                "tags": [
                    "comment"
                ]
            }
        },
        "/api/course/campuses": {
            "get": {
                "description": "根据课程名字获取课程开课校区",
                "parameters": [
                    {
                        "description": "课程名称关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetCourseCampusesResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取课程开课校区",
                "tags": [
                    "course"
                ]
            }
        },
        "/api/course/categories": {
            "get": {
                "description": "根据课程名字获取课程分类",
                "parameters": [
                    {
                        "description": "课程名称关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetCourseCategoriesResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取课程分类",
                "tags": [
                    "course"
                ]
            }
        },
        "/api/course/departs": {
            "get": {
                "description": "根据课程名字获取课程开课院系",
                "parameters": [
                    {
                        "description": "课程名称关键词",
                        "in": "query",
                        "name": "keyword",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetCourseDepartmentsResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取课程开课院系",
                "tags": [
                    "course"
                ]
            }
        },
        "/api/course/{courseId}": {
            "get": {
                "description": "获取课程信息",
                "parameters": [
                    {
                        "description": "课程ID",
                        "in": "path",
                        "name": "courseId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetCourseResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取课程信息",
                "tags": [
                    "course"
                ]
            }
        },
...
        "/api/search": {
            "post": {
                "description": "搜索课程列表",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/dto.ListCoursesReq",
                                        "summary": "body",
                                        "description": "ListCoursesReq"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "ListCoursesReq",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_ListCoursesResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "搜索课程列表",
                "tags": [
                    "course"
                ]
            }
        },
...
        "/api/search/total": {
            "get": {
                "description": "获取系统中所有课程评论的总数量",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.Response-dto_GetTotalCourseCommentsCountResp"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "summary": "获取吐槽总数",
                "tags": [
                    "comment"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}